finshed the meetings room
