+++ app/components/MeetingRoom.tsx
@@ -6,8 +6,16 @@ import {
   SpeakerLayout,
   useCallStateHooks,
 } from "@stream-io/video-react-sdk";
+import {
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
 
 import { CODING_QUESTIONS } from "@/constants";
+import { ScrollArea } from "@/components/ui/scroll-area";
 import { useRouter } from "next/navigation";
 import React, { useState } from "react";
 import {
@@ -24,7 +32,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { LayoutListIcon, UsersIcon } from "lucide-react";
 import Editor from "@monaco-editor/react";
-import QuestionsArea from "./QuestionsArea";
+import { Select } from "@radix-ui/react-select";
 
 const MeetingRoom = () => {
   const router = useRouter();
@@ -37,7 +45,17 @@ const MeetingRoom = () => {
 
   const callingState = useCallCallingState();
 
-  console.log(setLang, setQues);
+  function handleQuestionChange(question: string) {
+    const selected_ques = CODING_QUESTIONS.find((q) => q.id === question)!;
+    setQues(selected_ques);
+  }
+
+  //generated by CHATGPT
+  function handeLanguageChange(value: string) {
+    const newLang = value as "javascript" | "python" | "java";
+    setLang(newLang);
+    setCode(ques.starterCode[newLang]); // Update code when language changes
+  }
 
   if (callingState !== CallingState.JOINED) {
     return (
@@ -115,7 +133,106 @@ const MeetingRoom = () => {
         <ResizablePanel defaultSize={50}>
           <ResizablePanelGroup direction="vertical">
             <ResizablePanel defaultSize={25}>
-              <QuestionsArea />
+              <div className="">
+                <div className="flex items-center m-3 gap-7">
+                  <Select onValueChange={(value) => handeLanguageChange(value)}>
+                    <SelectTrigger className="w-[180px] border border-white">
+                      <SelectValue placeholder="Select a Language" />
+                    </SelectTrigger>
+                    <SelectContent>
+                      <SelectGroup>
+                        <SelectItem value="javascript">javascript</SelectItem>
+                        <SelectItem value="python">Python</SelectItem>
+                        <SelectItem value="java">Java</SelectItem>
+                      </SelectGroup>
+                    </SelectContent>
+                  </Select>
+                  <Select
+                    onValueChange={(value) => handleQuestionChange(value)}
+                  >
+                    <SelectTrigger className="w-[180px] border border-white">
+                      <SelectValue placeholder="Choose Question" />
+                    </SelectTrigger>
+                    <SelectContent>
+                      <SelectGroup>
+                        <SelectItem value="1">Question 1 (easy)</SelectItem>
+                        <SelectItem value="2">Question 2 (medium)</SelectItem>
+                        <SelectItem value="3">Question 3 (hard)</SelectItem>
+                      </SelectGroup>
+                    </SelectContent>
+                  </Select>
+                </div>
+
+                <ScrollArea className="h-[45vh] w-full rounded-md border flex flex-col mt-5 items-center p-10  border-purple-500 mr-5">
+                  <div className="">
+                    {ques.id === "1" ? (
+                      <>
+                        <h1 className="text-3xl font-bold mt-2 border-b-4 pb-2 border-purple-200">
+                          Question 1 (Easy) :- {ques.title}
+                        </h1>
+                        <p className="text-lg mt-10 border-[0.2px] rounded-lg border-purple-500 p-5">
+                          {ques.description}
+                        </p>
+
+                        {ques.examples.map((example, id) => (
+                          <div
+                            key={id}
+                            className="mt-10 border border-purple-400 p-5 rounded-lg"
+                          >
+                            <p>{example.input}</p>
+                            <p>{example.output}</p>
+                            <p>{example.explanation}</p>
+                          </div>
+                        ))}
+                      </>
+                    ) : ques.id === "2" ? (
+                      <>
+                        <>
+                          <h1 className="text-3xl font-bold mt-2 border-b-4 pb-2 border-purple-200">
+                            Question 2 (Medium) :- {ques.title}
+                          </h1>
+                          <p className="text-lg mt-10 border-[0.2px] rounded-lg border-purple-500 p-5">
+                            {ques.description}
+                          </p>
+
+                          {ques.examples.map((example, id) => (
+                            <div
+                              key={id}
+                              className="mt-10 border border-purple-400 p-5 rounded-lg"
+                            >
+                              <p>{example.input}</p>
+                              <p>{example.output}</p>
+                              <p>{example.explanation}</p>
+                            </div>
+                          ))}
+                        </>
+                      </>
+                    ) : (
+                      <>
+                        <>
+                          <h1 className="text-3xl font-bold mt-2 border-b-4 pb-2 border-purple-200">
+                            Question 3 (Hard) :- {ques.title}
+                          </h1>
+                          <p className="text-lg mt-10 border-[0.2px] rounded-lg border-purple-500 p-5">
+                            {ques.description}
+                          </p>
+
+                          {ques.examples.map((example, id) => (
+                            <div
+                              key={id}
+                              className="mt-10 border border-purple-400 p-5 rounded-lg"
+                            >
+                              <p>{example.input}</p>
+                              <p>{example.output}</p>
+                              <p>{example.explanation}</p>
+                            </div>
+                          ))}
+                        </>
+                      </>
+                    )}
+                  </div>
+                </ScrollArea>
+              </div>
             </ResizablePanel>
             <ResizableHandle />
             <ResizablePanel defaultSize={75}>
